Fix precision warning

+++ src/Compression/tests/gtest_compressionCodec.cpp
@@ -749,7 +749,7 @@ MonotonicGenerator() -> MonotonicGenerator<Int32>;
 auto RandomishGenerator = [](auto i)
 {
     using T = decltype(i);
-    double sin_value = sin(static_cast<double>(i * i)) * i;
+    double sin_value = sin(static_cast<double>(i * i)) * static_cast<double>(i);
     if (sin_value < std::numeric_limits<T>::lowest() || sin_value > static_cast<double>(std::numeric_limits<T>::max()))
         return T{};
     return T(sin_value);
